# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoolerControl Daemon API",
    "summary": "CoolerControl Rest Endpoints",
    "description": "Basic OpenAPI documentation for the CoolerControl Daemon API",
    "contact": {
      "name": "CoolerControl",
      "url": "https://coolercontrol.org"
    },
    "license": {
      "name": "GPL3+",
      "identifier": "GPL3+"
    },
    "version": "4.3.1"
  },
  "paths": {
...
    "/sse": {
      "get": {
        "tags": [
          "sse"
        ],
        "summary": "Server Sent Events",
        "description": "Subscribes to every event stream on one connection. Optional `events` query parameter narrows the subscription to a comma-separated subset of: status, health, logs, modes, alerts, notifications. Prefer this over the per-stream endpoints: browsers cap concurrent connections per origin over HTTP/1.1.",
        "parameters": [
          {
            "in": "query",
            "name": "events",
            "description": "Comma-separated substreams to subscribe to. All of them when absent.",
            "schema": {
              "description": "Comma-separated substreams to subscribe to. All of them when absent.",
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request. The request is invalid"
          },
          "401": {
            "description": "Unauthorized. Invalid credentials were provided."
          },
          "403": {
            "description": "Forbidden. Insufficient permissions."
          },
          "404": {
            "description": "Whatever you're looking for, it's not here."
          },
          "500": {
            "description": "An internal error has occurred."
          },
          "429": {
            "description": "Too Many Requests. Login attempts have been rate limited."
          },
          "502": {
            "description": "Bad Gateway. An error has occurred with an external library."
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/sse/status": {
      "get": {
        "tags": [
          "sse"
        ],
        "summary": "Recent Status Server Sent Events",
        "description": "Deprecated: use GET /sse?events=status,health. Subscribes and returns the Server Sent Events for a Status stream",
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cc",
        "description": "The private session cookie used for authentication."
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic",
        "bearerFormat": "",
        "description": "HTTP Basic authentication, mostly used to generate a secure authentication cookie."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "cc_<uuid>",
        "description": "Bearer token authentication for external services."
      }
    },
...
  },
...
}